Imported the remaining tabs from main
- Imported Account, Admin, Bookings, Home and index.
